Better Abc Tools multi-threading for auto exporter. As there is a log for GC to do when its running on many files, the UI popups can still become somewhat sluggish. Increased internal midi resolution for some songs for better divisibility. Unified abc exporter's use of song start and end ticks, it really was a mess. Allow for not being so strict with ignoring silence at start to first bar with notes, to prevent possible exception. Improved various toString() methods, and added some comments. Added some asserts for debugging.

+++ src/com/digero/common/midi/MidiUtils.java
@@ -109,13 +109,13 @@ public class MidiUtils {
 
 		// optimization to not always go through entire list of tempo events
 		int snapshotIndex = cache.snapshotIndex;
-		int snapshotMicro = cache.snapshotMicro;
+		long snapshotMicro = cache.snapshotMicro;
 
 		// walk through all tempo changes and add time for the respective blocks
-		long us = 0; // microsecond
+		long us = 0L; // microsecond
 
 		if (snapshotIndex <= 0 || snapshotIndex >= cacheCount || ticks[snapshotIndex] > tick) {
-			snapshotMicro = 0;
+			snapshotMicro = 0L;
 			snapshotIndex = 0;
 		}
 		if (cacheCount > 0) {
@@ -209,7 +209,7 @@ public class MidiUtils {
 		// index in ticks/tempos at the snapshot
 		int snapshotIndex = 0;
 		// microsecond at the snapshot
-		int snapshotMicro = 0;
+		long snapshotMicro = 0L;
 
 		int currTempo; // MPQ, used as return value for microsecond2tick
 
@@ -221,7 +221,7 @@ public class MidiUtils {
 			tempos = new int[1];
 			tempos[0] = DEFAULT_TEMPO_MPQ;
 			snapshotIndex = 0;
-			snapshotMicro = 0;
+			snapshotMicro = 0L;
 		}
 
 		public TempoCache(Sequence seq) {
@@ -267,7 +267,7 @@ public class MidiUtils {
 				tempos[e] = getTempoMPQ(evt.getMessage());
 			}
 			snapshotIndex = 0;
-			snapshotMicro = 0;
+			snapshotMicro = 0L;
 		}
 
 		public int getCurrTempoMPQ() {
